refactor: Car 클래스 Model 패키지로 이동
