+++ src/test/kotlin/racingcar/CarTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.model.Car
 
 class CarTest {
     @Test
