+++ src/test/kotlin/racingcar/ViewTest.kt
@@ -3,6 +3,7 @@ package racingcar
 import camp.nextstep.edu.missionutils.test.NsTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.model.Car
 
 class ViewTest : NsTest() {
     private val view = View()
